docker-edge-node/knowledge-base/edge/requirements.txt
requests==2.31.0
neo4j==5.14.1
python-dotenv==1.0.0